Report the flake8 version in .travis.yml (#9)
flake8 wraps several tools, and will run all of the ones that it knows about (and are installed). However, if a tool isn't installed, flake8 just silently doesn't run it. Fortunately, calling `flake8 --version` prints out the version of flake8 and all of the other tools it invokes.

In this case, that's: `3.6.0 (flake8-docstrings: 1.3.0, pydocstyle: 3.0.0, import-order: 0.18, mccabe: 0.6.1, naming: 0.7.0, pycodestyle: 2.4.0, pyflakes: 2.0.0) CPython 3.6.3 on Linux`

By calling `flake8 --version` every Travis run, issues should be easier to understand in the future.