Fix PyPI package and deployment
* Adds MANIFEST.in
The PyPI package cannot be installed because LICENSE and README.md are
missing from the source distribution (and are accessed by setup.py). We
add a MANIFEST.in file to tell distutils which non-module files should
be included in the distribution.

* Separates deployment into its own TravisCI build stage
By default, TravisCI runs all build stages for all configurations. This
means that the job for each Python version will attempt to deploy to
PyPI. We are creating a source distribution, so our Python version is
not part of the package spec, and all jobs will generate the same
package version and attempt to upload it to PyPI. PyPI only allows an
upload to happen once for each unique version, so one build job is
guaranteed to fail on every deploy. We address this by creating a
separate build job for deployment, which executes only once, after all
test jobs have successfully completed.