Automatically retrieve package version from tag
During deployment, we'd like to automatically set the package version
to the tag name from GitHub. Here we check the TRAVIS_TAG environment
variable in setup.py, and use whatever string is there if it exists. If
there is no tag, setup.py falls back to a DEFAULT_VERSION string, which
should be updated each release.

There are a few fancy packages for doing this automatically (e.g.
`versioneer` and `pbr`), but using `pbr` requires rewriting our setup
files, and versioneer forces you to commit autogenerated code to the
repo.

This setup is dead simple, unlikely to fail, and easy to understand.

Other modifications in this change:
* Set `long_description_content_type` to 'text/markdown' so that PyPI
  properly parses the README.md
* Change the `license` string to 'MIT', because setuptools only
  supports simple strings in this field. We originally read in the
  entire LICENSE file to this field, which causes errors.