Fix automated documentation generation (#25)
My automated versioning PR accidentally used a viskit path for the
version script, which broke docs builds and ReadTheDocs.

See https://readthedocs.org/projects/akro/builds/8677752/

This PR uses the correct path to the `__version__.py` script. It also
adds documentation generation to the CI, to detect failures like this
in the future.