Fix Dict.contains (#29)
Dict.contains was testing for membership of a set of spaces in a Dict
space, rather than membership of a sample in a Dict space.

This broke the build for newer versions of gym, though actually because
we used `contains` in a pickling test, and assumed a different behavior.

This PR fixes out version of `Dict.contains` so the tests will pass, it
does not add tests for contains, since we will soon replace our
implementations of basic space functionality with those inherited from
gym. See #28 for more info.